config: remove all tokens and add MOR token with its faucet
