+++ hardhat/erc20-token/README.md
@@ -1,16 +1,13 @@
 # Hardhat Project
 
-This project contains two contracts:
+This project contains one contract:
 
-1. `USDCredTokenWithFaucet.sol` - USD Credit ERC20 token with a faucet, to be deployed on the eosevm_testnet.
-2. `USDCredToken.sol` - USD Credit standard ERC20 token, to be deployed on the eosevm_mainnet.
-3. `MBOTSPLToken.sol` - MBOTS Credit ERC20 token, to be deployed on the eosevm_testnet and eosevm_mainnet.
+1. `MORTokenWithFaucet.sol` - MOR ERC20 token with a faucet, to be deployed on the Arbitrum Sepolia.
 
 ## Scripts
 
-- `deployUSDCredWithFaucet.js` - Script to deploy the USDCredTokenWithFaucet contract on eosevm_testnet.
-- `deployUSDCredWithoutFaucet.js` - Script to deploy the USDCredToken contract on eosevm_mainnet.
-- `deployMBOTSPLTestnet.js` - Script to deploy the USDCredToken contract on eosevm_testnet.
+- `deployMORWithFaucet.js` - Script to deploy the MORCredTokenWithFaucet contract on Arbitrum Sepolia.
+
 
 ## Network Configuration
 
@@ -24,9 +21,10 @@ PK=<your-private-key>
 
 To deploy the contracts, run the corresponding package.json script
 
+## How to deploy the contract
+
+1. Arbitrum Sepolia: run `npx hardhat run scripts/deployMORWithFaucet.js --network arbitrumSepolia`
+2. Arbitrum One: run `npx hardhat run scripts/deployMORWithFaucet.js --network arbitrumOne`
+
 ## Addresses
 
-### EOS EVM Testnet
-
-- USD Credit Token [0x1d5A4C37e60cAd0C72c057E3c191352429cDB38e](https://explorer.testnet.evm.eosnetwork.com/address/0x1d5A4C37e60cAd0C72c057E3c191352429cDB38e)
-- MBOTS Prelaunch Token [0x357752b66961021524b44523cD90a8B3861803E5](https://explorer.testnet.evm.eosnetwork.com/address/0x357752b66961021524b44523cD90a8B3861803E5)
