--- hardhat/erc20-token/scripts/deployUSDTWithFaucet.js
@@ -1,16 +1,0 @@
-
-const { ethers } = require("hardhat");
-
-async function main() {
-    const USDCredTokenWithFaucet = await ethers.getContractFactory("USDTTokenWithFaucet");
-    const token = await USDCredTokenWithFaucet.deploy(ethers.utils.parseUnits("1000000", 6));
-
-    await token.deployed();
-
-    console.log("USDTTokenWithFaucet deployed to:", token.address);
-}
-
-main().catch((error) => {
-    console.error(error);
-    process.exitCode = 1;
-});
